long double are changed in double since bambu does not support long double.

+++ panda_regressions/hls/gcc_regression_simple/memset-3.c
@@ -22,7 +22,7 @@
 static union {
   char buf[MAX_LENGTH];
   long long align_int;
-  long double align_fp;
+  double align_fp;
 } u;
 
 char A = 'A';
